app/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "kitchen-scheduler-backend"
version = "0.1.0"
description = "FastAPI backend for the kitchen scheduling system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.3",
    "alembic>=1.13.1",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.4.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"kitchen_scheduler.services.data" = ["*.json"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]
extend-exclude = [
    "build",
    "dist",
    ".venv",
    "venv",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I"]
ignore = ["E501"]
